requirements.txt
streamlit
numpy
pydub
numpy-stl
toml
scipy
pytest
matplotlib
